backend/requirements.txt
Django>=4.0
djangorestframework>=3.12
python-dateutil
